requirements.txt
streamlit
ultralytics
torch
Pillow
opencv-python-headless
